Prescriptions and MAR: Disallow users to perform edits on discharged encounters (#2133)
Prescriptions and MAR: Disallow users to perform edits on discharged encounters (#2133)

---------
